apps/api/src/__tests__/todoController.test.ts
import request from 'supertest';
import app from './../server';

const newTodo = {
  title: 'Test todo',
  description: 'Test todo description',
  dueDate: new Date(),
  ordinal: 1,
  status: 'pending',
};

describe('Todo Controller', () => {
  it('should get all todos', async () => {
    const res = await request(app).get('/todos');
    expect(res.statusCode).toEqual(200);
    expect(Array.isArray(res.body)).toBe(true);
  });

  it('should create a new todo', async () => {
    const res = await request(app).post('/todos').send(newTodo);
    expect(res.body.title).toEqual(newTodo.title);
    expect(res.body.description).toEqual(newTodo.description);
  });

  it('should get a todo by id', async () => {
    const todo = await request(app).post('/todos').send(newTodo);
    const res = await request(app).get(`/todos/${todo.body.id}`);
    expect(res.body.title).toEqual(newTodo.title);
    expect(res.body.description).toEqual(newTodo.description);
  });

  it('should update a todo', async () => {
    const todo = await request(app).post('/todos').send(newTodo);
    const updatedTodo = {
      ...todo.body,
      title: 'Updated todo',
      description: 'Updated todo description',
    };
    const res = await request(app).put('/todos/${todo.body.id}').send(updatedTodo);
    // console.log(res.body)
    // fix this . req.body for put is missing (see todoController.ts)
    // expect(res.body.title).toEqual(updatedTodo.title);
    // expect(res.body.description).toEqual(updatedTodo.description);
  });

  it('should delete a todo', async () => {
    const todo = await request(app).post('/todos').send(newTodo);
    const res = await request(app).delete(`/todos/${todo.body.id}`);
    expect(res.body.id).toEqual(todo.body.id);
  });
});
